refactor(LC041): split scalar projection analysis helpers
- separate query-chain analysis and local-usage analysis
- keep analyzer behavior unchanged while reducing local complexity
